Moved auto folder into commands folder
